requirements.txt
Django==1.7.4
Django-Select2==4.2.2
Markdown==2.5.2
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.2
Unidecode==0.04.17
argparse==1.2.1
bbcode==1.0.20
beautifulsoup4==4.3.2
cmsplugin-filer==0.10.1
cmsplugin-zinnia==0.7
dj-database-url==0.3.0
django-admin-enhancer==1.0.0
django-annoying==0.8.0
django-app-namespace-template-loader==0.2
django-appconf==0.6
django-blog-zinnia==0.15.1
django-bootstrap-form==3.1
django-classy-tags==0.6.1
django-cms==3.0.9
django-common==0.1.51
django-contrib-comments==1.5
django-filer==0.9.9
django-meta==0.2.1
django-meta-mixin==0.1.1
django-mptt==0.6.1
django-parler==1.2.1
django-polymorphic==0.6.1
django-realtime==1.1
django-registration-redux==1.1
django-reversion==1.8.5
django-sekizai==0.8.1
django-tagging==0.3.4
django-taggit==0.12.2
django-taggit-autosuggest==0.2.6
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.5
djangocms-blog==0.3.1
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.1
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.999
pybb==0.1.10
pyparsing==2.0.3
pytils==0.3
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
wsgiref==0.1.2
zinnia-theme-bootstrap==0.4
